\name{autoemail}
\alias{autoemail}
\title{AutoEmail}
\description{
  Automatically email a file to an address
  using the perl program.
}
\usage{
autoemail(eadd, sfile, hnote = "Exam Results")
}
\arguments{
  \item{eadd}{Email address }
  \item{sfile}{file to be sent }
  \item{hnote}{subject line }
}
\details{
  This program will work well in Linux and Mac
  where Perl is installed - I am not sure about Windows.
  Creates a unix executable file, if perl is present.
}
\value{
  Side Effects.
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}
\note{
  Need to change the from desgnation.
  
}
\seealso{IDandEM}

\keyword{misc}
