\name{getKEY}
\alias{getKEY}
\title{Read Key output
}
\description{Read Key output
}
\usage{
getKEY(fn)
}
\arguments{
  \item{fn}{character string file name
}
}
\details{Reads in the file output of ProfessR
  and returns a vector of answers
}
\value{vector of correct answers
}

\author{
Jonathan M. Lees<jonathan.lees@unc.edu>
}


\seealso{version.exam, prep.solution
}

\keyword{misc}
