\name{prep.solution}
\alias{prep.solution}

\title{Prepare Solution Files}
\description{
 Prepare Latex Solution Files
}
\usage{
prep.solution(ofile)
}

\arguments{
  \item{ofile}{output file name}
}
\details{
  Prepares the Latex header for the solution files
}
\value{
 Side Effects
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{prep.exam}
\examples{
prep.solution("solfile")
}

\keyword{misc}
