\name{seebank}
\alias{seebank}
\title{Print out a bank of questions}
\description{
Prints out a bank of questions, one at a time
  
}
\usage{
seebank(QB)
}
\arguments{
  \item{QB}{ QuestinBank Structure}
}

\value{
 Side effects
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}


\examples{

data(QBANK1)
seebank(QBANK1)


}
\keyword{misc}
